Refactor: 팀 랭킹 페이지 컴포넌트 분리 #78
