--- src/components/rank/team/TeamNormalRankCard.tsx
@@ -1,60 +1,0 @@
-'use client';
-
-import Image from 'next/image';
-import { TeamWithPoints } from '@/types/rank.type';
-import UserTitle from '@/components/common/UserTitle';
-import { USER_TITLE_MODE } from '@/constants/mode.constants';
-import { useState } from 'react';
-import { CommonModal } from '@/components/common/CommonModal';
-import { TeamDetailModal } from './TeamDetailModalContents';
-
-type Props = {
-  team: TeamWithPoints;
-  rank: number;
-};
-
-export const NormalRankTeamCard = ({ team, rank }: Props) => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  return (
-    <>
-      <article
-        onClick={() => setIsModalOpen(true)}
-        className="border rounded-3xl p-4 shadow-md bg-gray-100 w-full h-24 flex items-center"
-      >
-        {/* 랭킹 표시 */}
-        <UserTitle mode={USER_TITLE_MODE.CARD_RANK}>
-          {rank}
-          {rank === 1 ? 'st' : rank === 2 ? 'nd' : rank === 3 ? 'rd' : 'th'}
-        </UserTitle>
-
-        {/* 사용자 이미지 */}
-        <figure className="w-16 h-16 bg-gray-300 rounded-lg flex items-center justify-center mr-4">
-          <Image
-            src="/images/default-emblem.png"
-            alt={team.teamName ?? 'Team'}
-            width={64}
-            height={64}
-            className="object-cover rounded-lg"
-          />
-        </figure>
-
-        {/* 사용자 정보 (레벨, 이름, 소개...) */}
-        <section className="flex-1">
-          <UserTitle mode={USER_TITLE_MODE.CARD_LEVEL}>Lv. 1</UserTitle>
-          <UserTitle mode={USER_TITLE_MODE.CARD_NAME}>
-            {team.teamName}
-          </UserTitle>
-          <p>
-            {team.memberCount}/{team.maxTeamSize}
-          </p>
-          <p className="text-gray-600 truncate">{team.teamBio}</p>
-        </section>
-      </article>
-
-      <CommonModal isOpen={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <TeamDetailModal team={team} onClose={() => setIsModalOpen(false)} />
-      </CommonModal>
-    </>
-  );
-};
